Move client's tranport to a separate package
