--- pkg/client/unversioned/transport_test.go
@@ -1,166 +1,0 @@
-/*
-Copyright 2014 The Kubernetes Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package unversioned
-
-import (
-	"net/http"
-	"testing"
-
-	"k8s.io/kubernetes/pkg/api/testapi"
-)
-
-type testRoundTripper struct {
-	Request  *http.Request
-	Response *http.Response
-	Err      error
-}
-
-func (rt *testRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
-	rt.Request = req
-	return rt.Response, rt.Err
-}
-
-func TestBearerAuthRoundTripper(t *testing.T) {
-	rt := &testRoundTripper{}
-	req := &http.Request{}
-	NewBearerAuthRoundTripper("test", rt).RoundTrip(req)
-	if rt.Request == nil {
-		t.Fatalf("unexpected nil request: %v", rt)
-	}
-	if rt.Request == req {
-		t.Fatalf("round tripper should have copied request object: %#v", rt.Request)
-	}
-	if rt.Request.Header.Get("Authorization") != "Bearer test" {
-		t.Errorf("unexpected authorization header: %#v", rt.Request)
-	}
-}
-
-func TestBasicAuthRoundTripper(t *testing.T) {
-	rt := &testRoundTripper{}
-	req := &http.Request{}
-	NewBasicAuthRoundTripper("user", "pass", rt).RoundTrip(req)
-	if rt.Request == nil {
-		t.Fatalf("unexpected nil request: %v", rt)
-	}
-	if rt.Request == req {
-		t.Fatalf("round tripper should have copied request object: %#v", rt.Request)
-	}
-	if user, pass, found := rt.Request.BasicAuth(); !found || user != "user" || pass != "pass" {
-		t.Errorf("unexpected authorization header: %#v", rt.Request)
-	}
-}
-
-func TestUserAgentRoundTripper(t *testing.T) {
-	rt := &testRoundTripper{}
-	req := &http.Request{
-		Header: make(http.Header),
-	}
-	req.Header.Set("User-Agent", "other")
-	NewUserAgentRoundTripper("test", rt).RoundTrip(req)
-	if rt.Request == nil {
-		t.Fatalf("unexpected nil request: %v", rt)
-	}
-	if rt.Request != req {
-		t.Fatalf("round tripper should not have copied request object: %#v", rt.Request)
-	}
-	if rt.Request.Header.Get("User-Agent") != "other" {
-		t.Errorf("unexpected user agent header: %#v", rt.Request)
-	}
-
-	req = &http.Request{}
-	NewUserAgentRoundTripper("test", rt).RoundTrip(req)
-	if rt.Request == nil {
-		t.Fatalf("unexpected nil request: %v", rt)
-	}
-	if rt.Request == req {
-		t.Fatalf("round tripper should have copied request object: %#v", rt.Request)
-	}
-	if rt.Request.Header.Get("User-Agent") != "test" {
-		t.Errorf("unexpected user agent header: %#v", rt.Request)
-	}
-}
-
-func TestTLSConfigKey(t *testing.T) {
-	// Make sure config fields that don't affect the tls config don't affect the cache key
-	identicalConfigurations := map[string]*Config{
-		"empty":          {},
-		"host":           {Host: "foo"},
-		"prefix":         {Prefix: "foo"},
-		"version":        {Version: "foo"},
-		"codec":          {Codec: testapi.Default.Codec()},
-		"basic":          {Username: "bob", Password: "password"},
-		"bearer":         {BearerToken: "token"},
-		"user agent":     {UserAgent: "useragent"},
-		"transport":      {Transport: http.DefaultTransport},
-		"wrap transport": {WrapTransport: func(http.RoundTripper) http.RoundTripper { return nil }},
-		"qps/burst":      {QPS: 1.0, Burst: 10},
-	}
-	for nameA, valueA := range identicalConfigurations {
-		for nameB, valueB := range identicalConfigurations {
-			keyA, err := tlsConfigKey(valueA)
-			if err != nil {
-				t.Errorf("Unexpected error for %q: %v", nameA, err)
-				continue
-			}
-			keyB, err := tlsConfigKey(valueB)
-			if err != nil {
-				t.Errorf("Unexpected error for %q: %v", nameB, err)
-				continue
-			}
-			if keyA != keyB {
-				t.Errorf("Expected identical cache keys for %q and %q, got:\n\t%s\n\t%s", nameA, nameB, keyA, keyB)
-				continue
-			}
-		}
-	}
-
-	// Make sure config fields that affect the tls config affect the cache key
-	uniqueConfigurations := map[string]*Config{
-		"no tls":                  {},
-		"insecure":                {Insecure: true},
-		"cadata 1":                {TLSClientConfig: TLSClientConfig{CAData: []byte{1}}},
-		"cadata 2":                {TLSClientConfig: TLSClientConfig{CAData: []byte{2}}},
-		"cert 1, key 1":           {TLSClientConfig: TLSClientConfig{CertData: []byte{1}, KeyData: []byte{1}}},
-		"cert 1, key 2":           {TLSClientConfig: TLSClientConfig{CertData: []byte{1}, KeyData: []byte{2}}},
-		"cert 2, key 1":           {TLSClientConfig: TLSClientConfig{CertData: []byte{2}, KeyData: []byte{1}}},
-		"cert 2, key 2":           {TLSClientConfig: TLSClientConfig{CertData: []byte{2}, KeyData: []byte{2}}},
-		"cadata 1, cert 1, key 1": {TLSClientConfig: TLSClientConfig{CAData: []byte{1}, CertData: []byte{1}, KeyData: []byte{1}}},
-	}
-	for nameA, valueA := range uniqueConfigurations {
-		for nameB, valueB := range uniqueConfigurations {
-			// Don't compare to ourselves
-			if nameA == nameB {
-				continue
-			}
-
-			keyA, err := tlsConfigKey(valueA)
-			if err != nil {
-				t.Errorf("Unexpected error for %q: %v", nameA, err)
-				continue
-			}
-			keyB, err := tlsConfigKey(valueB)
-			if err != nil {
-				t.Errorf("Unexpected error for %q: %v", nameB, err)
-				continue
-			}
-			if keyA == keyB {
-				t.Errorf("Expected unique cache keys for %q and %q, got:\n\t%s\n\t%s", nameA, nameB, keyA, keyB)
-				continue
-			}
-		}
-	}
-}
